Test failure location incorrectly reported in utest.
When a test case fails then the overall test failure location should be
set to the case failure location. This fails, particularly when the case
that fails is the last to be run. The failure is due to the fact that
report_failure can be called recursively but also uses global state
(something that in itself really needs to be changed in the furture) and
although the case failure location is preserved, once the test teardown
starts, it isn't restored. This change fixes that.